+++ sb3_contrib/ppo_recurrent/ppo_recurrent.py
@@ -239,7 +239,7 @@ class RecurrentPPO(OnPolicyAlgorithm):
                 # Convert to pytorch tensor or to TensorDict
                 obs_tensor = obs_as_tensor(self._last_obs, self.device)
                 episode_starts = th.tensor(self._last_episode_starts, dtype=th.float32, device=self.device)
-                actions, values, log_probs, lstm_states = self.policy.forward(obs_tensor, lstm_states, episode_starts)
+                actions, values, log_probs, lstm_states = self.policy(obs_tensor, lstm_states, episode_starts)
 
             actions = actions.cpu().numpy()