+++ docs/misc/changelog.rst
@@ -20,6 +20,7 @@ New Features:
 Bug Fixes:
 ^^^^^^^^^^
 - Fix RecurrentPPO and MaskablePPO forward and predict do not reshape action before clip it
+- Do not call ``forward()`` method directly in ``RecurrentPPO``
 
 Deprecations:
 ^^^^^^^^^^^^^